callist image and detil ui and loading

+++ app/src/main/java/com/nikhil/mybestfriend/feature/cat/data/db/entity/CatEntity.kt
@@ -14,5 +14,5 @@ data class CatEntity (
     val imperialWeight: String,
     val metricWeight: String,
     var url: String? = null,
-    val rating: Int
+    val rating: Double
 )
