+++ app/src/main/java/com/nikhil/mybestfriend/feature/cat/data/db/unitlocalized/MetricCatEntity.kt
@@ -16,9 +16,9 @@ data class MetricCatEntity (
     @ColumnInfo(name = "url")
     override var url: String?=null,
     @ColumnInfo(name = "rating")
-    override val rating: Int,
+    override val rating: Double,
     @ColumnInfo(name = "metricWeight")
     override val weight: String
 ) : UnitCatEntity{
-    constructor() : this("abys","null","null","null","null",null,0,"null")
+    constructor() : this("abys","null","null","null","null",null,0.0,"null")
 }
